mailboat.mta: refactored
