+++ mailboat/mta/__init__.py
@@ -1,101 +1,25 @@
 import asyncio
 import email.policy
+import logging
 from asyncio import Future
+from asyncio.locks import Lock
 from concurrent.futures import ThreadPoolExecutor
 from copy import deepcopy
-from dataclasses import dataclass
 from email.headerregistry import BaseHeader
 from email.message import EmailMessage
 from email.parser import Parser
-import logging
-from time import perf_counter
-from typing import Any, Awaitable, Callable, Dict, List, Optional, Protocol, Tuple, cast
-import aiosmtplib
-import aiosmtpd
+from typing import Dict, List, Optional, Tuple, cast
 
+import aiosmtplib
 from aiosmtpd.controller import Controller
-from aiosmtpd.handlers import AsyncMessage
-from aiosmtpd.smtp import AuthResult, Envelope, LoginPassword, SMTP, Session
+from aiosmtpd.smtp import AuthResult
 from aiosmtplib.errors import SMTPAuthenticationError
 from flanker.addresslib import address
 from unqlite import Collection, UnQLite
 
-from asyncio.locks import Lock
-
-from base64 import standard_b64decode, standard_b64encode
-
-from .utils.perf import async_perf_point
-
-LocalDeliveryHandler = Callable[[EmailMessage], Awaitable[Any]]
-SMTPAuthHandler = Callable[[SMTP, str, Any], Awaitable[AuthResult]] # the second parameter is method, the third is the data.
-# As method "login", "plain", the data is a `LoginPassword`
-
-class EmailQueue(Protocol):
-    def get(self) -> Awaitable[Tuple[EmailMessage, int]]:
-        ...
-
-    def remove(self, index: int) -> Awaitable[None]:
-        ...
-
-    def put(self, email: EmailMessage) -> Awaitable[None]:
-        ...
-
-
-@dataclass
-class DeliveryTask(object):
-    message_id: str
-    delivering_address: str
-    done: bool
-    success: bool
-    status_message: Optional[str] = None
-    tried: int = 0
-
-
-class _SMTPDHandler(AsyncMessage):  # TODO: support OAuth2
-    __logger = logging.getLogger("mailboat.mta._SMTPDHandler")
-
-    def __init__(
-        self,
-        message_handler: Callable[[EmailMessage], Awaitable[Any]],
-        smtp_auth_handler: SMTPAuthHandler,
-    ) -> None:
-        self.smtp_auth_handler = smtp_auth_handler
-        self.message_handler = message_handler
-        self.delivery_tasks: List[DeliveryTask] = []
-        super().__init__()
-
-    async def auth_LOGIN(self, server: SMTP, args: List[str]) -> AuthResult:
-        username = await server.challenge_auth("Username:")
-        if username == aiosmtpd.smtp.MISSING:
-            return AuthResult(success=False, handled=False)
-        else:
-            username = standard_b64decode(cast(bytes, username))
-        password = await server.challenge_auth("Password:")
-        if password == aiosmtpd.smtp.MISSING:
-            return AuthResult(success=False, handled=False)
-        else:
-            password = standard_b64decode(cast(bytes, password))
-        return await self.smtp_auth_handler(
-            server, "login", LoginPassword(cast(bytes, username), cast(bytes, password))
-        )
-
-    async def auth_PLAIN(self, server: SMTP, args: List[str]) -> AuthResult:
-        words = await server.challenge_auth("", encode_to_b64=False)
-        if words == aiosmtpd.smtp.MISSING:
-            return AuthResult(success=False, handled=False)
-        decoded_words = standard_b64decode(cast(bytes, words))
-        parts = decoded_words.split(b"\0")
-        if parts:
-            if not parts[0]:
-                parts.pop(0)
-        if len(parts) != 2:
-            return AuthResult(success=False, handled=True)
-        return await self.smtp_auth_handler(
-            server, "plain", LoginPassword(parts[0], parts[1])
-        )
-
-    async def handle_message(self, message: EmailMessage):
-        await self.message_handler(message)
+from ..utils.perf import async_perf_point
+from .protocols import EmailQueue, LocalDeliveryHandler, SMTPAuthHandler
+from .smtp import SMTPDHandler
 
 
 class UnQLiteEmailMessageQueue(EmailQueue):
@@ -113,7 +37,9 @@ class UnQLiteEmailMessageQueue(EmailQueue):
 
     async def get(self) -> Tuple[EmailMessage, int]:
         while len(self._ids) == 0:
-            await asyncio.sleep(0)  # TODO (rubicon): more effective way to block coroutine
+            await asyncio.sleep(
+                0
+            )  # TODO (rubicon): more effective way to block coroutine
         result = self._coll.fetch(self._ids.pop(0))
         doc_id: int = result["__id"]
         message: str = result["message"]
@@ -183,9 +109,8 @@ class TransferAgent(object):
                 database.collection("{}.queue".format(self_name))
             )
         )
-        self.delivery_tasks: List[DeliveryTask] = []
         self.smtpd_controller = Controller(
-            _SMTPDHandler(
+            SMTPDHandler(
                 self.handle_message,
                 smtp_auth_handler=smtpd_auth_handler,
             ),
@@ -270,43 +195,11 @@ class TransferAgent(object):
                         del message["bcc"]
                         message["bcc"] = message["delivered-to"]
                     delivered_to = address.parse(message["delivered-to"])
-                    delivery_task = DeliveryTask(
-                        message["message-id"],
-                        delivering_address=delivered_to.address,
-                        done=False,
-                        success=False,
-                    )
-                    self.delivery_tasks.append(
-                        delivery_task
-                    )
                     # TODO (rubicon): use a eventbus instead a list for delivery tasks
                     if delivered_to.hostname in self.mydomains:
-                        fut = asyncio.ensure_future(
-                            self.local_delivery_handler(message)
-                        )
-
-                        @fut.add_done_callback
-                        def when_local_delivery_done(fut: Future):
-                            asyncio.ensure_future(self.queue.remove(index))
-                            delivery_task.done = True
-                            if not fut.exception():
-                                delivery_task.success = True
-                            else:
-                                delivery_task.success = False
-                                delivery_task.status_message = str(fut.exception())
-
+                        asyncio.ensure_future(self.local_delivery_handler(message))
                     else:
-                        fut = asyncio.ensure_future(self.remote_deliver(message))
-
-                        @fut.add_done_callback
-                        def when_remote_delivery_done(fut: Future):
-                            asyncio.ensure_future(self.queue.remove(index))
-                            delivery_task.done = True
-                            if not fut.exception():
-                                delivery_task.success = True
-                            else:
-                                delivery_task.success = False
-                                delivery_task.status_message = str(fut.exception())
+                        asyncio.ensure_future(self.remote_deliver(message))
 
             except Exception as e:
                 __logger.exception(exc_info=e)
